pyproject.toml
[tool.poetry]
name = "web-django"
version = "0.1.0"
description = ""
authors = ["Nguyễn Đức Hoàng <hoang.nguyen@unified.com>"]
readme = "README.md"
packages = [{include = "djweb"}, {include = "render"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.6"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.1"
black = "^23.1.0"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
